// src/modules/lookups/lookup-kind.enum.ts
export enum LookupKind {
  Universities = 'universities',
  Companies = 'companies',
  Majors = 'majors',
  AcademicYears = 'academic-years',
  Students = 'students',
  Applications = 'applications',
  Opportunities = 'opportunities',
  StudentInternships = 'student-internships',
  UniversitySupervisors = 'university-supervisors',
  CompanySupervisors = 'company-supervisors',
}